requirements.txt
python-telegram-bot
python-dotenv
telethon
ollama
openai